fix: Re-implement S2L's view with React portal
* Due to overflow: 'hidden' set on Block to enforce border radius,
S2L's VisualCopy cannot show correctly. This fix use React portal
so that the VisualCopy does not need to be a descendant of
SearchTool, which is a Block, thus is not effected by
overflow: 'hidden'.

* This fix also upgrades the VisualCopy -- Use Block instead of a simple
custom wrapper, to make the VisualCopy look exactly the same as what
users will see after linking.